+++ lib/osam_common_module_flutter.dart
@@ -8,5 +8,7 @@ export 'src/model/language.dart';
 export 'src/model/rating_control_response.dart';
 export 'src/model/subscription_response.dart';
 export 'src/model/token_response.dart';
+export 'src/model/version.dart';
 export 'src/model/version_control_response.dart';
 export 'src/osam.dart';
+export 'src/ui/ui_helper.dart';